requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-groq
faiss-cpu
sentence-transformers
transformers
torch
python-dotenv
streamlit